+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ### Added
 
+- Can configure SSL verification and usage of SSL certs for API requests from frontend.
+
 ### Changed
 
 - Updated requests to version 2.32.0
 - Added startup commands to minhash and allele clustering services Dockerfiles.
+- Use pydantic-settings for config management in frontend.
 - Updated the samples tables to make sample name, sample id, and major spp searchable.
+
 
 ### Fixed